TST: annexrepo: Clear repo instances when testing shared connections
test_annex_ssh() goes through a convoluted dance to create an annex
repo, set up an SSH remote through GitRepo.add_remote(), and then get
an AnnexRepo instance.  The idea seems to be to add a remote without
using AnnexRepo.add_remote(), because that calls
_set_shared_connection(), and then check that AnnexRepo.__init__()
sets the shared connect.

However, __init__() isn't actually invoked because there's already a
known instance, and that's returned instead.  Clear the known instance
so that __init__() is triggered.